Refactor contentMetaData and add new tabs for challenges
