+++ test/sanityQueryService.test.js
@@ -39,6 +39,10 @@ const {
     FilterBuilder,
 } = require('../src/filterBuilder.js');
 
+const {
+    processMetadata,
+} = require('../src/contentMetaData.js');
+
 describe('Sanity Queries', function () {
     beforeEach(() => {
         const config = {
@@ -495,7 +499,7 @@ describe('Filter Builder', function () {
         expect(clauses[1].operator).toBe('>=');
     });
 
-    function getPlusUser() {
+    function getPlusUser() {challenges
         return {
             permissions: [91,92],
         }
@@ -517,4 +521,52 @@ describe('Filter Builder', function () {
         return clauses;
     }
 
+});
+
+describe('MetaData', function () {
+
+    test('custom', async () => {
+        const metaData = processMetadata('guitareo', 'recording');
+        expect(metaData.type).toBe('recording');
+        expect(metaData.name).toBe('Archives');
+        expect(metaData.description).toBeDefined();
+    });
+
+    test('onlyCommon', async () => {
+        const guitareoMetaData = processMetadata('guitareo', 'challenge');
+        const drumeoMetaData = processMetadata('drumeo', 'challenge');
+        expect(guitareoMetaData).toStrictEqual(drumeoMetaData);
+        expect(guitareoMetaData.type).toBe('challenge');
+        expect(guitareoMetaData.name).toBe('Challenges');
+    });
+
+    test('withCommon', async () => {
+        const guitareoMetaData = processMetadata('guitareo', 'coaches');
+        const drumeoMetaData = processMetadata('drumeo', 'coaches');
+        expect(guitareoMetaData.description).not.toBe(drumeoMetaData.description);
+        guitareoMetaData.description = ''
+        drumeoMetaData.description = ''
+        expect(guitareoMetaData).toStrictEqual(drumeoMetaData);
+    });
+
+    test('withWithoutFilters', async () => {
+        let metaData = processMetadata('singeo', 'student-review', true);
+        expect(metaData.type).toBeDefined()
+        expect(metaData.name).toBeDefined()
+        expect(metaData.description).toBeDefined();
+        expect(metaData.thumbnailUrl).toBeDefined();
+        expect(metaData.tabs).toBeDefined();
+        metaData = processMetadata('singeo', 'student-review', false);
+        expect(metaData.type).toBeDefined()
+        expect(metaData.name).toBeDefined()
+        expect(metaData.description).toBeDefined();
+        expect(metaData.tabs).not.toBeDefined();
+    });
+
+    test('nulled', async () => {
+        let metaData = processMetadata('drumeo', 'student-review');
+        expect(metaData).toBeNull();
+        metaData = processMetadata('singeo', 'student-review');
+        expect(metaData).not.toBeNull();
+    });
 });
